glob expansion for star, question mark, tilde

+++ src/eskilib/eskilib_file.c
@@ -1,12 +1,12 @@
 #include "eskilib_file.h"
 
 //simple fgets implementation that returns the number of characters read
-int_fast32_t eskilib_fgets(char* input_buffer, int size_of_input_buffer, FILE* file_pointer) {
+int_fast32_t eskilib_fgets(char* input_buffer, size_t size_of_input_buffer, FILE* file_pointer) {
 	register int character;
 	register char* buffer = input_buffer;
 	int_fast32_t characters_read = 0;
 
-	while(--size_of_input_buffer > 0 && (character = getc(file_pointer)) != EOF) {
+	while (--size_of_input_buffer > 0 && (character = getc(file_pointer)) != EOF) {
 		++characters_read;
 
 		if((*buffer = character) == '\n')
@@ -19,12 +19,12 @@ int_fast32_t eskilib_fgets(char* input_buffer, int size_of_input_buffer, FILE* f
 	return characters_read;
 }
 
-int_fast32_t eskilib_fgets_delimited(char* input_buffer, int size_of_input_buffer, FILE* file_pointer, char delimiter) {
+int_fast32_t eskilib_fgets_delimited(char* input_buffer, size_t size_of_input_buffer, FILE* file_pointer, char delimiter) {
 	register int character;
 	register char* buffer = input_buffer;
 	int_fast32_t characters_read = 0;
 
-	while(--size_of_input_buffer > 0 && (character = getc(file_pointer)) != EOF) {
+	while (--size_of_input_buffer > 0 && (character = getc(file_pointer)) != EOF) {
 		++characters_read;
 
 		if((*buffer = character) == delimiter)
