+++ src/ncsh_history.h
@@ -36,7 +36,7 @@ void ncsh_history_clean(struct ncsh_History* history);
 
 void ncsh_history_free(struct ncsh_History* history);
 
-enum eskilib_Result ncsh_history_add(char* line, uint_fast32_t length, struct ncsh_History* history);
+enum eskilib_Result ncsh_history_add(char* line, size_t length, struct ncsh_History* history);
 
 // enum eskilib_Result ncsh_history_remove(char* line, uint_fast32_t length, struct ncsh_History* history);
