+++ src/ncsh_parser.h
@@ -3,9 +3,9 @@
 #ifndef ncsh_parser_h
 #define ncsh_parser_h
 
-#include <stdint.h>
+#include <stddef.h>
 
-void ncsh_parse(char line[], uint_fast32_t length, struct ncsh_Args* args);
+void ncsh_parse(char line[], size_t length, struct ncsh_Args* args);
 
 #endif // !ncsh_parser_h
